requirements.txt
Flask==3.1.0
google-cloud-datastore==2.20.1